Refactor to remove extension methods and directly inject them instead
Remove all other warnings
Add new project to pull the UI code into so it can be tested properly and shared

+++ YoutubeCleanupTool.Domain/GetAndCacheYouTubeData.cs
@@ -135,7 +135,7 @@ namespace YouTubeCleanupTool.Domain
             var videoTitles = await context.GetVideoTitles();
             foreach (var videoTitle in videoTitles)
             {
-                if (videoTitle.ToCharArray().Any(x => ((int)x) > 1000))
+                if (videoTitle.ToCharArray().Any(x => x > 1000))
                 {
                     callback(videoTitle);
                 }
